requirements.txt
paho-mqtt
pandas
plotly
streamlit-extras